Rename to Snowplow Elasticsearch Loader (closes #17)
