+++ README.md
@@ -1,41 +1,40 @@
-# Kinesis Elasticsearch Sink
+# Snowplow Elasticsearch Loader
 
 ## Introduction
 
-The Kinesis Elasticsearch Sink consumes Snowplow enriched events or failed events from an [Amazon Kinesis][kinesis] stream, transforms them to JSON, and writes them to [Elasticsearch][elasticsearch]. Events which cannot be transformed or which are rejected by Elasticsearch are written to a separate Kinesis stream.
+The Snowplow Elasticsearch Loader consumes Snowplow enriched events or failed events from an
+[Amazon Kinesis][kinesis] stream, transforms them to JSON, and writes them to
+[Elasticsearch][elasticsearch]. Events which cannot be transformed or which are rejected by
+Elasticsearch are written to a separate Kinesis stream.
 
 ## Building
 
-Assuming you already have [SBT 0.13.0][sbt] installed:
+Assuming you already have [SBT 0.13][sbt] installed:
 
-    $ git clone git://github.com/snowplow/snowplow.git
-    $ cd 4-storage/kinesis-elasticsearch-sink
+    $ git clone git://github.com/snowplow/snowplow-elasticsearch-loader.git
     $ sbt compile
-    
+
 ## Usage
 
-The Kinesis Elasticsearch Sink has the following command-line interface:
+The Snowplow Elasticsearch Loader has the following command-line interface:
 
 ```
-snowplow-elasticsearch-sink: Version 0.1.0. Copyright (c) 2014, Snowplow Analytics
-Ltd.
+snowplow-elasticsearch-loader 0.8.0
 
-Usage: snowplow-elasticsearch-sink [OPTIONS]
+Usage: snowplow-elasticsearch-loader [options]
 
-OPTIONS
---config filename
-                   Configuration file.
+  --config <filename>
 ```
 
 ## Running
 
 Create your own config file:
 
-    $ cp src/main/resources/application.conf.example my.conf
+    $ cp examples/config.hocon.sample my.conf
 
-Edit it and update the AWS credentials:
+Update the configuration to fit your needs like modifying the AWS credentials:
 
-```js
+```json
 aws {
   access-key: "default"
   secret-key: "default"
@@ -44,18 +43,18 @@ aws {
 
 Next, start the sink, making sure to specify your new config file:
 
-    $ sbt "run --config my.conf"
+    $ ./snowplow-elasticsearch-loader-http-0.8.0 --config my.conf
 
 ## Find out more
 
-| Technical Docs              | Setup Guide           | Roadmap & Contributing               |         
+| Technical Docs              | Setup Guide           | Roadmap & Contributing               |
 |-----------------------------|-----------------------|--------------------------------------|
-| ![i1][techdocs-image]      | ![i2][setup-image]   | ![i3][roadmap-image]                |
-| [Technical Docs][techdocs] | [Setup Guide][setup] | _coming soon_                        |
+| ![i1][techdocs-image]       | ![i2][setup-image]    | ![i3][roadmap-image]                 |
+| [Technical Docs][techdocs]  | [Setup Guide][setup]  | _coming soon_                        |
 
 ## Copyright and license
 
-Copyright 2014 Snowplow Analytics Ltd.
+Copyright 2014-2017 Snowplow Analytics Ltd.
 
 Licensed under the [Apache License, Version 2.0][license] (the "License");
 you may not use this software except in compliance with the License.
@@ -69,7 +68,7 @@ limitations under the License.
 [kinesis]: http://aws.amazon.com/kinesis/
 [snowplow]: http://snowplowanalytics.com
 [elasticsearch]: http://www.elasticsearch.org/
-[sbt]: http://typesafe.artifactoryonline.com/typesafe/ivy-releases/org.scala-sbt/sbt-launch/0.13.0/sbt-launch.jar
+[sbt]: http://www.scala-sbt.org
 
 [setup]: https://github.com/snowplow/snowplow/wiki/kinesis-elasticsearch-sink-setup
 [techdocs]: https://github.com/snowplow/snowplow/wiki/kinesis-elasticsearch-sink
