+++ core/src/main/scala/com.snowplowanalytics.elasticsearch/loader/CredentialsLookup.scala
@@ -17,7 +17,7 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
 // Amazon
 import com.amazonaws.auth._
