+++ core/src/main/scala/com.snowplowanalytics.elasticsearch/loader/ElasticsearchSinkExecutor.scala
@@ -17,11 +17,10 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow
-package storage.kinesis.elasticsearch
+package com.snowplowanalytics
+package elasticsearch.loader
 
 // AWS Kinesis Connector libs
-import com.amazonaws.services.kinesis.connectors.elasticsearch.ElasticsearchObject
 import com.amazonaws.services.kinesis.connectors.{
   KinesisConnectorConfiguration,
   KinesisConnectorExecutorBase,
@@ -29,7 +28,7 @@ import com.amazonaws.services.kinesis.connectors.{
 }
 
 // Tracker
-import scalatracker.Tracker
+import snowplow.scalatracker.Tracker
 
 // This project
 import clients.ElasticsearchSender
