+++ core/src/main/scala/com.snowplowanalytics.elasticsearch/loader/JsonRecord.scala
@@ -17,11 +17,12 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
 /**
- * Like the connector library's ITransformer, but for consuming from stdin
+ * Format in which Snowplow events are buffered
+ *
+ * @param json The JSON string for the event
+ * @param id The event ID (if it exists)
  */
-trait StdinTransformer {
-  def consumeLine(line: String): EmitterInput
-}
+case class JsonRecord(json: String, id: Option[String])
