+++ core/src/main/scala/com.snowplowanalytics.elasticsearch/loader/Shredder.scala
@@ -17,7 +17,7 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
 // Scalaz
 import scalaz._
@@ -38,7 +38,7 @@ import scala.annotation.tailrec
  */
 object Shredder {
 
-  private[elasticsearch] val schemaPattern =
+  private[loader] val schemaPattern =
     """^iglu:([a-zA-Z0-9-_.]+)/([a-zA-Z0-9-_]+)/[a-zA-Z0-9-_]+/([0-9]+-[0-9]+-[0-9]+)$""".r
 
   /**
