+++ core/src/main/scala/com.snowplowanalytics.elasticsearch/loader/StdinTransformer.scala
@@ -17,12 +17,9 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
-/**
- * Format in which Snowplow events are buffered
- *
- * @param json The JSON string for the event
- * @param id The event ID (if it exists)
- */
-case class JsonRecord(json: String, id: Option[String])
+/** Like the connector library's ITransformer, but for consuming from stdin */
+trait StdinTransformer {
+  def consumeLine(line: String): EmitterInput
+}
