+++ core/src/test/scala/com.snowplowanalytics.elasticsearch.loader/BadEventTransformerSpec.scala
@@ -11,16 +11,11 @@
  * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow
-package storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
 // Scalaz
 import scalaz._
 import Scalaz._
-
-// json4s
-import org.json4s._
-import org.json4s.jackson.JsonMethods._
 
 // Specs2
 import org.specs2.mutable.Specification
