+++ http/src/main/scala/com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch/ElasticsearchHTTPSinkApp.scala
@@ -17,7 +17,7 @@
  * governing permissions and limitations there under.
  */
 
-package com.snowplowanalytics.snowplow.storage.kinesis.elasticsearch
+package com.snowplowanalytics.elasticsearch.loader
 
 import clients.{ElasticsearchSender, ElasticsearchSenderHTTP}
